Added localstack and more tests
