+++ .github/workflows/pipeline.yml
@@ -14,3 +14,5 @@ permissions:
 jobs:
   test-python:
     uses: NERC-CEH/dri-cicd/.github/workflows/test-python.yml@main
+    with:
+      localstack_profile_name: localstack
